Creating tests for swapimodel

+++ KneatSoftware/KneatSoftware/Interfaces/ISwapiApiResponse.cs
@@ -1,12 +1,10 @@
 ﻿using KneatSoftware.Models;
-using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace KneatSoftware.Interfaces
 {
     public interface ISwapiApiResponse
     {
-        Dictionary<string, int> StartCalculateTotalOfStops(IList<SwapiApiResponse> swapiApiResponses, int megalightsTyped);
+        Dictionary<string, int> StartCalculateTotalOfStops(IList<SwapiApiResponse> swapiApiResponses, double megalightsTyped);
     }
 }
